A relatively minimal but slightly haphazard code shuffle to fix dependencies.txt.
