+++ icu4c/source/test/depstest/dependencies.txt
@@ -869,7 +869,8 @@ library: i18n
     dayperiodrules
     listformatter
     formatting formattable_cnv regex regex_cnv translit
-    double_conversion number_representation number_output numberformatter number_skeletons numberparser
+    double_conversion number_representation number_output numberformatter
+    number_skeletons numberformatter_microprops numberparser
     units_extra unitsformatter
     universal_time_scale
     uclean_i18n
@@ -989,16 +990,30 @@ group: numberformatter
     number_decimfmtprops.o
     number_fluent.o number_formatimpl.o
     number_grouping.o number_integerwidth.o number_longnames.o
-    number_mapper.o number_modifiers.o number_multiplier.o
+    number_mapper.o number_modifiers.o
     number_notation.o number_padding.o
-    number_patternmodifier.o number_patternstring.o number_rounding.o
-    number_scientific.o number_usageprefs.o
+    number_patternmodifier.o number_patternstring.o
+    number_scientific.o
     currpinf.o dcfmtsym.o numsys.o
     numrange_fluent.o numrange_impl.o
   deps
     decnumber double_conversion formattable units unitsformatter
     number_representation number_output
+    numberformatter_microprops
     uclean_i18n common
+
+group: numberformatter_microprops
+    number_multiplier.o 
+    number_usageprefs.o
+  deps
+    number_rounding
+    unitsformatter
+
+group: number_rounding
+    number_rounding.o
+  deps
+    currency
+    number_representation
 
 group: number_skeletons
     # Number skeleton support; separated from numberformatter
